Replace bindings with Bindgen-Cs Generated

+++ tests/native/Inno.Native.Sdl3.Tests/Sdl3InitTests.cs
@@ -15,7 +15,7 @@ public sealed class Sdl3InitTests
     [Fact]
     public void InitAndQuit_ShouldSucceed()
     {
-        var initResult = SDL.Init((uint)SDLInitFlags.Events);
+        var initResult = SDL.Init(SDLInitFlags.Events);
         if (!initResult)
         {
             var error = SDL.GetErrorAsException();
